+++ lib/core/theme/cubit/theme_provider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:kaban_frontend/core/theme/domain/cubit/theme_bloc.dart';
+import 'package:kaban_frontend/core/theme/cubit/theme_bloc.dart';
 
 class ThemeProvider extends StatelessWidget {
   final Widget child;
