+++ lib/main.dart
@@ -3,9 +3,9 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:kaban_frontend/core/domain/entity/env_type.dart';
 import 'package:kaban_frontend/core/router/app_router.dart';
 import 'package:kaban_frontend/core/config/bloc/config_provider.dart';
-import 'package:kaban_frontend/core/theme/domain/cubit/theme_bloc.dart';
-import 'package:kaban_frontend/core/theme/domain/cubit/theme_provider.dart';
-import 'package:kaban_frontend/core/theme/domain/cubit/theme_state.dart';
+import 'package:kaban_frontend/core/theme/cubit/theme_bloc.dart';
+import 'package:kaban_frontend/core/theme/cubit/theme_provider.dart';
+import 'package:kaban_frontend/core/theme/cubit/theme_state.dart';
 
 void main() {
   runApp(AppWidget());
